requirements.txt
pyspark
sqlalchemy
panda
psycopg2-binary
python-dotenv
psutil
